version bump 1.0.0
Hooray since this and previous release candidates have been running in production for awhile I think it is time to call this 1.0.

I am proud of what I have been able to prove out with this gem, but not to proud of hwo the code reads at this time. I will likely make some refactorings and cleanup on the 1.X branch trying to not change the api to much. After getting things cleaned up a bit, and offering some needed features like dynamic real time configuration option settings. I will likely make some api changes to get a cleaner interaction with the client, and try to go 2.X at that point... but that is a ways off and this should be good to use in various projects. Help me improve the docs, and make this gem easier to use.