Initial work to drop Simplecov and Simplecov HTML as dependencies
This is a start to build out full support for a mountable Rack UI which can support additional features beyond the orignal static reports.

* Showing the time range the coverage is valid for
* Ability to filter coverage to a specific time range
* Improved UI and navigation
* Gem coverage usage summaries
* etc

This pulls the initial views, assets (JS, CSS) from Simple-Cov HTML as well as some files to do calculations on coverage data into the project. Initially with little changes.