fix for issue #188 broken file merge over deploy names...
The pervious merge functionality was tested at to low of a level and missed that with additional meta data. The merge was broken and regressed a few releases ago. This test scrubbed real world data captures and resolves the issue

Thanks to @gingerlime for the great bug report and data and support to figure this out.