handle stdlib branch coverage options, now a default in simplecov 0.18.x we will get more bug reports for projects that use both
This handles two issues:

* we need to be able to parse the new formats in stdlib Ruby Coverage 2.5 & 2.6, namely :lines which comes along with branching support, since SimpleCov 0.18.x the :lines format is a default
* also, Simplecov doesn't check if Coverage is already running, and we start Coverage first which causes an error if both aree configured. I detect that case and output a warning, as Coverage will be started late and miss some files our eager_loading hook

The first fix is really valuable as anyone can configure the various Coverage format options, the simplecov specific detecting is a bit annoying and specific to if both libs are trying to run in test mode. For now we don't really recommended running Coverband in test.