Fix specs for Ruby 3.1
Ruby 3.1 has made some [breaking changes](https://bugs.ruby-lang.org/issues/18176] to Coverage API, which causes Simplecov to not work well with Coverband together.

This PR skips Simplecov for any Ruby >= 3.1 as a temporary fix.

This PR also patches `HashRedisStore` for using ActiveSupport's `blank?` method, which somehow does not get loaded in Ruby 3.1. This is not a proper solution but I do think it would be nice to not use any ActiveSupport/Rails APIs unless necessary.