Introduce send_deferred_eager_loading_data configuration to reduce the load on Redis during deployment
This is another way to avoid cache stampede by omitting some reporting on starting servers.
Coverband stores the results of eager_loading to Redis at server startup.
The eager_loading results are big and the same for all servers, so there is no need to save all results.
By configuring the eager_loading results of some servers to be stored in Redis,
we can reduce the load on Redis during deployment.